Refactored the management of "subjective" models.
* Conceptually, the `accounting` application is orthogonal to the
other ones in our project, since the facilities it provides can be
useful within any context where objects as `Account`, `Transaction` or
`Invoice` make sense.  In fact, in the current implementation, it
relies only on the generic `Subject` model, which can be mapped to
domain-specific "subjective" models (e.g. GAS, Person, Supplier).

So, in order to make the application more generic, we need to decouple
it from other apps.  In this respect, is more convenient to implement
the `Subject` model as an "interface" model (using the `ContentType`
framework) rather than a mix-in class.

* Also added some basic configuration for the admin site

to
be
implemented as an "interface" model than 

it relies only on a generic Subject
should be